[Feat] on k8s mode get checkpoint support.

+++ streamx-plugin/streamx-flink-kubernetes/src/main/scala/com/streamxhub/streamx/flink/kubernetes/model/CheckpointCV.scala
@@ -19,10 +19,8 @@
 
 package com.streamxhub.streamx.flink.kubernetes.model
 
-case class CheckpointCV(
-                         id: Integer,
-                         checkpointPath: String,
-                         isSavepoint: Boolean,
-                         checkpointType: String,
-                         pollAckTime: Long
-                       )
+case class CheckpointCV(id: Long,
+                        checkpointPath: String,
+                        isSavepoint: Boolean,
+                        checkpointType: String,
+                        pollAckTime: Long)
